fixed order quantity input and fixed commercial_registration_number for farmer and supplier

+++ database/migrations/2024_08_08_095121_create_suppliers_table.php
@@ -19,12 +19,11 @@ return new class extends Migration
             $table->string('image')->nullable();
             $table->unsignedBigInteger('owner_id');
             $table->string('type');
-            $table->integer('commercial_registration_number');
+            $table->unsignedBigInteger('commercial_registration_number');
             $table->boolean('status')->default(true);
             $table->softDeletes();
             $table->timestamps();
             $table->foreign('owner_id')->references('id')->on('users');
-
         });
     }
